fix: make hog schema migration non-breaking for prod deploy
The roaming-hogs change redefined the `hog` table that #29 had already
published to prod (static {x,y} i32). Changing x/y to f64 and adding
dir/movedAt columns was a destructive migration, so the production
deploy aborted (the workflow refuses --delete-data by design).

Keep x/y as i32 and give the new dirX/dirY/movedAt columns defaults, so
the change is an in-place additive migration. A Hog now re-bases at a
whole tile each wander tick (clients still glide between via
projectMotion) — it never pushes, so it needs no sub-tile precision.

Verified by publishing #29's schema to a local db and migrating this
module over it: applies in place, Hogs seed and roam on integer tiles.

Generated-By: PostHog Code
Task-Id: 8f685706-29a9-4e9b-8583-d17fc841e701

+++ src/module_bindings/hog_table.ts
@@ -13,8 +13,8 @@ import {
 export default __t.row({
   id: __t.u64().primaryKey(),
   zoneId: __t.string().name("zone_id"),
-  x: __t.f64(),
-  y: __t.f64(),
+  x: __t.i32(),
+  y: __t.i32(),
   dirX: __t.i32().name("dir_x"),
   dirY: __t.i32().name("dir_y"),
   movedAt: __t.timestamp().name("moved_at"),
